Fix issue #14, and general cleanup
Swagger.Core has a painful setup of the `Swagger` model, expecting it
to be configured on the `ServletContext` (which is ugly and annoying to
do in dropwizard)
I’ve added a convenience constructor to pass in a default `Swagger`
model.

Also cleaned up a lot of the documentation, and removed the
`SwagerUIConfigurator.forConfig` methods, which i found to be more
annoying then useful, and moved them to just be straight constructors.